backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
google-auth==2.25.2
cryptography==41.0.7
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
sentence-transformers==2.2.2
python-docx==1.1.0
torch==2.9.0
numpy==1.26.2
requests==2.31.0
